Notice refactor -- Move the NoticeExporter interface definition to the notice package
