+++ domain/src/main/java/org/mobilitydata/gtfsvalidator/domain/entity/notice/error/FloatFieldValueOutOfRangeNotice.java
@@ -16,7 +16,7 @@
 
 package org.mobilitydata.gtfsvalidator.domain.entity.notice.error;
 
-import org.mobilitydata.gtfsvalidator.domain.entity.NoticeExporter;
+import org.mobilitydata.gtfsvalidator.domain.entity.notice.NoticeExporter;
 import org.mobilitydata.gtfsvalidator.domain.entity.notice.base.ErrorNotice;
 
 import java.io.IOException;
